:truck: api: move find-finalized-sessions route and controller from lib to src
