+++ api/lib/application/sessions/index.js
@@ -71,33 +71,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/admin/sessions/to-publish',
-      config: {
-        validate: {
-          query: Joi.object({
-            filter: {
-              version: Joi.number(),
-            },
-          }),
-        },
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: finalizedSessionController.findFinalizedSessionsToPublish,
-        tags: ['api', 'finalized-sessions'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/admin/sessions/with-required-action',
       config: {
         validate: {
