+++ api/src/certification/session-management/domain/usecases/find-finalized-sessions-to-publish.js
