+++ api/tests/certification/session-management/unit/domain/usecases/find-finalized-sessions-to-publish_test.js
@@ -1,5 +1,5 @@
-import { findFinalizedSessionsToPublish } from '../../../../lib/domain/usecases/find-finalized-sessions-to-publish.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { findFinalizedSessionsToPublish } from '../../../../../../src/certification/session-management/domain/usecases/find-finalized-sessions-to-publish.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | findFinalizedSessionsToPublish', function () {
   let finalizedSessionRepository;
